implementando area do triangulo e criando arquivos

+++ include/area.h
@@ -3,6 +3,6 @@
 
 #include <iostream>
 
-int AreaTriangulo (float base, float altura);
+float AreaTriangulo (float base, float altura);
 
 #endif
