+++ src/area.cpp
@@ -4,6 +4,6 @@
 using std::cin;
 using std::cout;
 
-int AreaTriangulo (float base, float altura) {
+float AreaTriangulo (float base, float altura) {
 	return (base * altura)/2.0;
 }
